added follows constraint
added the follows constraint.
this allows us to specify that a course must be *immediately* followed by another (specific) course.
this is used to handle e.g. 4hr courses where there's a separate lab in a different timeslot (e.g. 1170 and 2170).

this required a lot of changes to the recursive solving algorithm and the pre-solve prep work.
however, with clever adjustments to the course topology structure we now get this constraint for free due to new pset iteration order guarantees.
this means that finding a solution with this constraint is no more difficult than without (perhaps even easier due to shrinking the search space).

updated the demo course file with follows constraints