added schedule constraint
added the schedule constraint for the courses file.
this allows you to specify a specific schedule that the course must go into.

rewrote the iteration structure again: now performed in terms of follow pset chains.
this means we don't even need the visited set anymore.
this works because all the follow sets are in the order in which they will need to invoke each other recursively already.

parser for courses file now takes the list of schedules to sanitize specification of schedule ids.