refactor
refactored the project organization to be more conducive to scalability.
things that used to be list searched are now separately hashed for O(1) lookup.
all constraint info is now specified in a single file (thus avoiding confusing definition order).
if said file is not specified, reads from stdin (wasn't possible before because separate files).
all parsing logic is now delegated to handler functions which are looked up in a hash table.
constraint instructions are now case-insensitive.