max adj courses constraint and optimizations
added constraint for instructors to enforce a max number of adjacent courses.
this is specified inline for all instructors (not a separate constraint).
this prevents an instructor from having e.g. 3+ courses to teach back-to-back with no time in between for prep.

fixed a few mistakes in the constraints file regarding room/course capacities as they currently stand based on S2020 schedule.

created the index_union type that maintains info about contiguous indexes.
this is now additionally maintained during backtracking search in order to enforce the max adjacent course constraints.
this is maintained in the pset-level of backtrack search.
care must be taken to undo changes on backtracking, but comments are added to warn about this before and after the dangerous code block (currently correct).
preventing instructors from teaching two different courses at same time is now done in terms of the new index unions for each timeslot (O(1) rather than O(n) lookup).

fixed a few places where the parsing logic was allowing formats that technically worked but weren't intended.

rewrote satisfy_info (which performs searches) to store a consumer function for satisfied solutions.
an enum class named yield_t is returned from this consumer.
if `cont` is returned, search continues.
if `stop` is returned, search ends.
currently the passed function just prints the solution and requests to stop searching.
later on this may be used for fancier things like selecting a set of solutions.

added a bunch of schedule-based info for searching.
this is now rolled into a struct and mapped via schedule_to_info.
this absorbs the old schedule_to_follow_chain_to_start_index (way too long anyway).
additions are assigned instructors and assigned courses for each timeslot in each schedule.
these are both maintained in the course-level of backtrack search.

typedefed the follow chain type as fchain_t.
this shortens up a lot of things (esp. search functions, which now fit on one screen).

orthogonality checks are now done by looping through ortho courses for each course, rather than the other way around.
this allows us to make use of the new assigned_courses field for each timeslot in each schedule and avoid a lot of work.
this is especially faster as the number of rooms increases (now O(1) rather than O(n)).